+++ posthog-core/src/types.ts
@@ -71,6 +71,7 @@ export enum PostHogPersistedProperty {
   Queue = 'queue',
   OptedOut = 'opted_out',
   SessionId = 'session_id',
+  SessionStartTimestamp = 'session_start_timestamp',
   SessionLastTimestamp = 'session_timestamp',
   PersonProperties = 'person_properties',
   GroupProperties = 'group_properties',
@@ -115,7 +116,11 @@ export type PostHogQueueItem = {
 }
 
 export type PostHogEventProperties = {
-  [key: string]: any
+  [key: string]: JsonType
+}
+
+export type PostHogGroupProperties = {
+  [type: string]: string | number
 }
 
 export type PostHogAutocaptureElement = {
@@ -125,9 +130,8 @@ export type PostHogAutocaptureElement = {
   nth_child?: number
   nth_of_type?: number
   order?: number
-} & {
-  [key: string]: any
-} // Any key prefixed with `attr__` can be added
+} & PostHogEventProperties
+// Any key prefixed with `attr__` can be added
 
 export type PostHogRemoteConfig = {
   sessionRecording?:
@@ -236,7 +240,7 @@ export type PostHogFlagsAndPayloadsResponse = Partial<
   Pick<PostHogFlagsResponse, 'featureFlags' | 'featureFlagPayloads'>
 >
 
-export type JsonType = string | number | boolean | null | { [key: string]: JsonType } | Array<JsonType>
+export type JsonType = string | number | boolean | null | { [key: string]: JsonType } | Array<JsonType> | JsonType[]
 
 export type FetchLike = (url: string, options: PostHogFetchOptions) => Promise<PostHogFetchResponse>